+++ cli/src/printer.rs
@@ -1,6 +1,6 @@
 use colored::Colorize;
 use prettytable::{format, row, Row, Table};
-use reinfer_client::{Bucket, Dataset, Project, Source, Stream, User};
+use reinfer_client::{resources::quota::Quota, Bucket, Dataset, Project, Source, Stream, User};
 use serde::{Serialize, Serializer};
 
 use anyhow::{anyhow, Context, Error, Result};
@@ -72,6 +72,16 @@ impl DisplayTable for Bucket {
                 None => "missing".dimmed(),
             }
         ]
+    }
+}
+
+impl DisplayTable for Quota {
+    fn to_table_headers() -> Row {
+        row![bFg => "Kind", "Hard Limit"]
+    }
+
+    fn to_table_row(&self) -> Row {
+        row![self.quota_kind, self.hard_limit]
     }
 }